+++ svc/TrackService.ts
@@ -119,6 +119,7 @@ export class TrackService extends HoistService {
             msg: stripTags(options.message),
             clientUsername: XH.getUsername(),
             appVersion: XH.getEnv('clientVersion'),
+            clientAppCode: XH.clientAppCode,
             loadId: XH.loadId,
             tabId: XH.tabId,
             url: window.location.href,