+++ svc/WebSocketService.ts
@@ -45,7 +45,13 @@ export class WebSocketService extends HoistService {
     readonly REG_SUCCESS_TOPIC = 'xhRegistrationSuccess';
     readonly FORCE_APP_SUSPEND_TOPIC = 'xhForceAppSuspend';
     readonly REQ_CLIENT_HEALTH_RPT_TOPIC = 'xhRequestClientHealthReport';
-    readonly METADATA_FOR_HANDSHAKE = ['appVersion', 'appBuild', 'loadId', 'tabId'];
+    readonly METADATA_FOR_HANDSHAKE = [
+        'appVersion',
+        'appBuild',
+        'loadId',
+        'tabId',
+        'clientAppCode'
+    ];
 
     /** True if WebSockets not explicitly disabled via {@link AppSpec.disableWebSockets}. */
     enabled: boolean = !XH.appSpec.disableWebSockets;